backend/requirements.txt
# requirements.txt - Python Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
websockets==12.0
aiosqlite==0.19.0
httpx==0.25.1
redis==5.0.1
celery==5.3.4
flower==2.0.1
pytz==2023.3.post1
qrcode==7.4.2
Pillow==10.1.0

# OpenAI
openai==1.3.0

# Telegram (if needed)
python-telegram-bot==20.3
aiogram==2.25.1

# Additional utilities
aiofiles==23.2.1
pandas==2.1.4
numpy==1.24.3
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Monitoring
prometheus-client==0.19.0
structlog==23.2.0

# Security
bcrypt==4.1.2
cryptography==41.0.7
aiohttp==3.9.1  # Async HTTP client